메인 DB부분 수정

+++ src/main/java/bitc/full502/sceneshare/domain/entity/dto/MovieInfoDTO.java
@@ -9,6 +9,5 @@ public class MovieInfoDTO {
     private String movieTitle;
     private Integer movieRatingAvg;
     private String moviePosterUrl;
-    private int bookmarkCnt;
-
+    private Long bookmarkCnt;
 }
